Refactor: Split filters module & Code style
* Split filters side panel view into multiple sub-modules.
* Check code style in the modules improving the code without logic
  changes.
* Remove low value unit tests
